ia_service/requirements.txt
fastapi
uvicorn
pillow
opencv-python
numpy
pytesseract
pandas
ta-lib 